+++ tests/test_wq/test_wq_wel.py
@@ -4,6 +4,7 @@ import shutil
 import numpy as np
 import pandas as pd
 import pytest
+import xarray as xr
 
 from imod.wq import Well
 
@@ -24,6 +25,20 @@ def well_conc():
     return Well(
         id_name="well", x=x, y=y, rate=5.0, layer=2, time=datetimes, concentration=2.5
     )
+
+
+@pytest.fixture(scope="module")
+def well_conc_multiple_species():
+    datetimes = pd.date_range("2000-01-01", "2000-01-05")
+    y = np.arange(4.5, 0.0, -1.0)
+    x = np.arange(0.5, 5.0, 1.0)
+    w = Well(
+        id_name="well", x=x, y=y, rate=5.0, layer=2, time=datetimes, concentration=2.5
+    )
+    conc1 = w["concentration"].assign_coords(species=1)
+    conc2 = w["concentration"].assign_coords(species=2)
+    w["concentration"] = xr.concat([conc1, conc2], dim="species")
+    return w
 
 
 def test_render(well):
@@ -173,6 +188,37 @@ def test_render_concentration(well_conc, tmp_path):
     assert actual == compare
 
 
+def test_render_concentration_multiple_species(well_conc_multiple_species, tmp_path):
+    wel = well_conc_multiple_species
+    directory = pathlib.Path("well")
+    # Test rate, not affected by multiple species
+    compare = """
+    wel_p1_s1_l2 = well/well_20000101000000_l2.ipf
+    wel_p2_s1_l2 = well/well_20000102000000_l2.ipf
+    wel_p3_s1_l2 = well/well_20000103000000_l2.ipf
+    wel_p4_s1_l2 = well/well_20000104000000_l2.ipf
+    wel_p5_s1_l2 = well/well_20000105000000_l2.ipf"""
+
+    actual = wel._render(directory, globaltimes=wel["time"].values, system_index=1)
+    assert actual == compare
+
+    # Test concentration
+    compare = """
+    cwel_t1_p1_l2 = well/well-concentration_c1_20000101000000_l2.ipf
+    cwel_t1_p2_l2 = well/well-concentration_c1_20000102000000_l2.ipf
+    cwel_t1_p3_l2 = well/well-concentration_c1_20000103000000_l2.ipf
+    cwel_t1_p4_l2 = well/well-concentration_c1_20000104000000_l2.ipf
+    cwel_t1_p5_l2 = well/well-concentration_c1_20000105000000_l2.ipf
+    cwel_t2_p1_l2 = well/well-concentration_c2_20000101000000_l2.ipf
+    cwel_t2_p2_l2 = well/well-concentration_c2_20000102000000_l2.ipf
+    cwel_t2_p3_l2 = well/well-concentration_c2_20000103000000_l2.ipf
+    cwel_t2_p4_l2 = well/well-concentration_c2_20000104000000_l2.ipf
+    cwel_t2_p5_l2 = well/well-concentration_c2_20000105000000_l2.ipf"""
+
+    actual = wel._render_ssm(directory, globaltimes=wel["time"].values)
+    assert actual == compare
+
+
 def test_save(well, tmp_path):
     wel = well
     wel.save(tmp_path / "well")
@@ -203,6 +249,31 @@ def test_save(well_conc, tmp_path):
         "well-concentration_20000103000000_l2.ipf",
         "well-concentration_20000104000000_l2.ipf",
         "well-concentration_20000105000000_l2.ipf",
+    ]
+    for file in files:
+        assert (tmp_path / "well" / file).is_file()
+
+
+def test_save_multiple_species(well_conc_multiple_species, tmp_path):
+    wel = well_conc_multiple_species
+    wel.save(tmp_path / "well")
+
+    files = [
+        "well_20000101000000_l2.ipf",
+        "well_20000102000000_l2.ipf",
+        "well_20000103000000_l2.ipf",
+        "well_20000104000000_l2.ipf",
+        "well_20000105000000_l2.ipf",
+        "well-concentration_c1_20000101000000_l2.ipf",
+        "well-concentration_c1_20000102000000_l2.ipf",
+        "well-concentration_c1_20000103000000_l2.ipf",
+        "well-concentration_c1_20000104000000_l2.ipf",
+        "well-concentration_c1_20000105000000_l2.ipf",
+        "well-concentration_c2_20000101000000_l2.ipf",
+        "well-concentration_c2_20000102000000_l2.ipf",
+        "well-concentration_c2_20000103000000_l2.ipf",
+        "well-concentration_c2_20000104000000_l2.ipf",
+        "well-concentration_c2_20000105000000_l2.ipf",
     ]
     for file in files:
         assert (tmp_path / "well" / file).is_file()